Crate polkadot_sdk_docs_first_runtime

Runtime used in your_first_runtime.

Modules

api

genesis_config_presets

Provides getters for genesis configuration presets.

interface

Just a handy re-definition of some types based on what is already provided to the pallet configs.

Structs

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

Enums

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

RuntimeViewFunction

Runtime query type.

Constants

RUNTIME_API_VERSIONS

VERSION

Type Aliases

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Balances

BalancesConfig

FirstPallet

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

TransactionPaymentConfig